✨ 이미지 캐싱 로직, 이미지 서버 url load 로직 통합 및 적용
