+++ StreetDrop/StreetDrop/Presentation/MainScene/View/DroppedMusicWithinAreaCollectionViewCell.swift
@@ -13,6 +13,7 @@ import RxSwift
 
 final class DroppedMusicWithinAreaCollectionViewCell: UICollectionViewCell {
     static let identifier = "DroppedMusicWithinAreaCollectionViewCell"
+    private let disposeBag: DisposeBag = DisposeBag()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -26,7 +27,7 @@ final class DroppedMusicWithinAreaCollectionViewCell: UICollectionViewCell {
     func setData(item: MusicWithinAreaEntity) {
         self.musicTitleLabel.text = item.musicTitle
         self.singerNameLabel.text = item.artist
-        self.albumCoverImageView.setImage(with: item.albumImageURL)
+        self.albumCoverImageView.setImage(with: item.albumImageURL, disposeBag: disposeBag)
     }
     
     // MARK: - UI
